+++ e2e/cases/css/tailwindcss-mjs/postcss.config.mjs
@@ -1,0 +1,12 @@
+import { fileURLToPath } from 'url';
+import { join, dirname } from 'path';
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+export default {
+  plugins: {
+    tailwindcss: {
+      config: join(__dirname, './tailwind.config.mjs'),
+    },
+  },
+};